requirements.txt
psycopg2-binary==2.8.6
pydantic==1.8.1
python-dotenv==0.17.1